+++ core/src/websocket/aio/client.rs
@@ -1045,6 +1045,7 @@ impl WebSocketClient {
                     Arc::clone(&subscriptions),
                     Arc::clone(&shutdown_requested),
                     Arc::clone(&reconnection),
+                    Arc::clone(&state),
                 )
                 .await;